+++ build.gradle
@@ -84,51 +84,53 @@ repositories {
   maven { url 'http://repo.spring.io/milestone' }
   maven { url 'https://oss.jfrog.org/libs-snapshot' }
   maven { url 'https://dl.bintray.com/reactivesocket/RSocket' }
+  maven { url "https://oss.sonatype.org/content/repositories/releases" }
   maven { url "https://jitpack.io" }
   mavenCentral()
+  jcenter()
 }
 
 dependencies {
-  compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:${kotlin_version}"
-  compile "org.jetbrains.kotlin:kotlin-reflect:${kotlin_version}"
+  implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${kotlin_version}"
+//  implementation "org.jetbrains.kotlin:kotlin-reflect:${kotlin_version}"
+  implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:0.22.5'
 
-  compile 'io.rsocket:rsocket-transport-netty:0.10.3'
-  compile 'io.rsocket:rsocket-transport-local:0.10.3'
+  implementation 'io.rsocket:rsocket-transport-netty:0.10.3'
+  implementation 'io.rsocket:rsocket-transport-local:0.10.3'
 
-  compile 'io.airlift:airline:0.8'
-  compile 'com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.9.4'
-  compile 'com.fasterxml.jackson.core:jackson-core:2.9.4'
-  compile 'com.fasterxml.jackson.core:jackson-databind:2.9.4'
-  compile 'com.fasterxml.jackson.module:jackson-module-afterburner:2.9.4'
-  compile 'com.fasterxml.jackson.dataformat:jackson-dataformat-cbor:2.9.4'
-  compile 'org.slf4j:slf4j-jdk14:1.7.25'
+  implementation 'io.airlift:airline:0.8'
+  implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.9.4'
+  implementation 'com.fasterxml.jackson.core:jackson-core:2.9.4'
+  implementation 'com.fasterxml.jackson.core:jackson-databind:2.9.4'
+  implementation 'com.fasterxml.jackson.module:jackson-module-afterburner:2.9.4'
+  implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-cbor:2.9.4'
+  implementation 'org.slf4j:slf4j-jdk14:1.7.25'
 
-  compile 'org.eclipse.jetty.http2:http2-http-client-transport:9.4.8.v20171121'
-  compile 'org.mortbay.jetty.alpn:alpn-boot:8.1.12.v20180117'
-  compile 'com.github.vjames19.kotlin-futures:kotlin-futures-jdk8:0.4.0'
-  compile 'com.spotify:completable-futures:0.3.2'
-  compile 'com.spotify:completable-futures:0.3.2'
+  implementation 'org.eclipse.jetty.http2:http2-http-client-transport:9.4.9.v20180320'
+  implementation 'org.mortbay.jetty.alpn:alpn-boot:8.1.12.v20180117'
+//  implementation 'com.github.vjames19.kotlin-futures:kotlin-futures-jdk8:0.4.0'
+  implementation 'com.spotify:completable-futures:0.3.2'
 
-  runtime 'org.slf4j:slf4j-jdk14:1.7.25'
+  implementation 'com.baulsupp:oksocial-output:1.8.0'
+
+  implementation 'org.slf4j:slf4j-jdk14:1.8.0-beta2'
 
   // JUnit Jupiter API and TestEngine implementation
-  testCompile("org.junit.jupiter:junit-jupiter-api:${junitJupiterVersion}")
-  testRuntime("org.junit.jupiter:junit-jupiter-engine:${junitJupiterVersion}")
+  testCompile "org.junit.jupiter:junit-jupiter-api:${junitJupiterVersion}"
+  testRuntime "org.junit.jupiter:junit-jupiter-engine:${junitJupiterVersion}"
 
   // If you also want to support JUnit 3 and JUnit 4 tests
-  testRuntime("org.junit.vintage:junit-vintage-engine:5.1.0")
+  testRuntime "org.junit.vintage:junit-vintage-engine:5.1.0"
 
-  testRuntime("org.junit.platform:junit-platform-launcher:1.1.0")
+  testRuntime "org.junit.platform:junit-platform-launcher:1.1.0"
 
   testCompile "org.jetbrains.kotlin:kotlin-test:${kotlin_version}"
   testCompile "org.jetbrains.kotlin:kotlin-test-junit:${kotlin_version}"
 
-  testRuntime group: 'org.slf4j', name: 'slf4j-jdk14', version: '1.7.25'
+  testRuntime 'org.slf4j:slf4j-jdk14:1.8.0-beta2'
 
-  testCompile group: 'com.squareup.okhttp3', name: 'okhttp-testing-support', version: '3.10.0'
-  testCompile group: 'com.squareup.okhttp3', name: 'mockwebserver', version: '3.10.0'
-
-  compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:${kotlin_version}"
+  testCompile 'com.squareup.okhttp3:okhttp-testing-support:3.10.0'
+  testCompile 'com.squareup.okhttp3:mockwebserver:3.10.0'
 }
 
 test {
@@ -154,5 +156,11 @@ test {
     }
 
     failFast = true
+}
+
+kotlin {
+  experimental {
+    coroutines "enable"
+  }
 }
