--- src/main/kotlin/io/rsocket/cli/XxX.kt
@@ -1,1 +1,0 @@
-package io.rsocket.cli
