+++ src/main/kotlin/io/rsocket/cli/http2/Http2ClientTransport.kt
@@ -19,15 +19,15 @@ import java.net.URI
 import java.util.function.Supplier
 
 class Http2ClientTransport @JvmOverloads constructor(
-        private val uri: URI,
-        private var transportHeadersFn: () -> Map<String, String> = { mutableMapOf() }) : ClientTransport, TransportHeaderAware {
+  private val uri: URI,
+  private var transportHeadersFn: () -> Map<String, String> = { mutableMapOf() }) : ClientTransport, TransportHeaderAware {
 
   override fun setTransportHeaders(transportHeaders: Supplier<Map<String, String>>?) {
     this.transportHeadersFn = { transportHeaders?.get() ?: mapOf() }
   }
 
   override fun connect(): Mono<DuplexConnection> =
-          createSession().flatMap { s -> Http2DuplexConnection.create(s, uri, transportHeadersFn()) }
+    createSession().flatMap { s -> Http2DuplexConnection.create(s, uri, transportHeadersFn()) }
 
   private fun createSession(): Mono<Session> = Mono.create { c ->
     val client = HTTP2Client()
@@ -43,25 +43,25 @@ class Http2ClientTransport @JvmOverloads constructor(
       client.start()
 
       client.connect(
-              sslContextFactory,
-              InetSocketAddress(uri.host, port),
-              ServerSessionListener.Adapter(),
-              object : Promise<Session> {
-                override fun succeeded(result: Session?) {
-                  c.success(result)
-                }
+        sslContextFactory,
+        InetSocketAddress(uri.host, port),
+        ServerSessionListener.Adapter(),
+        object : Promise<Session> {
+          override fun succeeded(result: Session?) {
+            c.success(result)
+          }
 
-                override fun failed(x: Throwable?) {
-                  c.error(x!!)
-                }
-              })
+          override fun failed(x: Throwable?) {
+            c.error(x!!)
+          }
+        })
     } catch (e: Exception) {
       c.error(e)
     }
   }
 
   private fun daemonClientScheduler(): ScheduledExecutorScheduler =
-          ScheduledExecutorScheduler("jetty-scheduler", true)
+    ScheduledExecutorScheduler("jetty-scheduler", true)
 
   private fun daemonClientExecutor(): QueuedThreadPool {
     val executor = QueuedThreadPool()
