+++ src/main/kotlin/io/rsocket/cli/http2/Http2UriHandler.kt
@@ -9,8 +9,8 @@ import java.util.Optional.of
 
 class Http2UriHandler : UriHandler {
   override fun buildClient(uri: URI?): Optional<ClientTransport> =
-          when {
-            HttpScheme.HTTPS.`is`(uri!!.scheme) || HttpScheme.HTTP.`is`(uri.scheme) -> of(Http2ClientTransport(uri))
-            else -> super.buildClient(uri)
-          }
+    when {
+      HttpScheme.HTTPS.`is`(uri!!.scheme) || HttpScheme.HTTP.`is`(uri.scheme) -> of(Http2ClientTransport(uri))
+      else -> super.buildClient(uri)
+    }
 }
